Added <picture> element w/ media and srcset attributes
Also edited the src attributes of all images to match the new image names and paths. NOTE: Not all images were placed inside a <picture> element as it was not deemed necessary at this time. In those cases, only srcset was used to give the browser the option between the 2x or 1x version of an image.