HTML & CSS clean-up - first pass
HTML: 1) removed outermost <div class="container"> because it was not necessary to achieve flexbox layout. 2) removed the height attribute from <img class="project-image"> and added it to the CSS under the project-image class.

CSS: 1) Improved the responsiveness of the header's border-bottom. 2) Cleaned-up the overall styles by consolidating and removing redundancies and uncessary styles (that were not being applied, as I found by looking at DevTools)